requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django==5.2
djangorestframework==3.16.0
mypy-extensions==1.0.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
ruff==0.11.6
sqlparse==0.5.3
